chore(Moengage): move sdk loader to different file
